+++ src/core/transaction.ts
@@ -1,9 +1,9 @@
+import { Signer } from '@ethersproject/abstract-signer';
+import { Wallet } from '@ethersproject/wallet';
+import { SignedTx } from '@vocdoni/proto/vochain';
+import { Buffer } from 'buffer';
 import { strip0x } from '../util/common';
 import { Signing } from '../util/signing';
-import { Wallet } from '@ethersproject/wallet';
-import { SignedTx } from '../dvote-protobuf/build/ts/vochain/vochain';
-import { Signer } from '@ethersproject/abstract-signer';
-import { Buffer } from 'buffer';
 
 export abstract class TransactionCore {
   /**